+++ include/thodd/lang/pod.hpp
@@ -344,58 +344,6 @@ thodd::lang
     
 
 
-    template <
-        typename lang_t,
-        typename iterator_t>
-    struct token
-    {
-        std::pair<iterator_t, iterator_t> data ;
-        lang_t id ;
-
-        inline auto 
-        begin () 
-        {
-            return data.first ;
-        }
-
-        inline auto const
-        begin () const 
-        {
-            return data.first ;
-        }
-
-        inline auto 
-        end () 
-        {
-            return data.second ;
-        }
-
-        inline auto const
-        end () const
-        {
-            return data.second ;
-        }
-
-        inline auto const
-        size() const 
-        {
-            return 
-            std::distance(begin(), end()) ;
-        }
-
-        inline auto const
-        invalid () const
-        {
-            return 
-            size() == 0 ;
-        }
-    } ;
-
-
-    template<
-        typename lang_t, 
-        typename iterator_t>
-    token (std::pair<iterator_t, iterator_t> const &, lang_t) -> token<lang_t, iterator_t> ;
 
 
 
@@ -404,135 +352,10 @@ thodd::lang
 
 
 
-
-
-
-
-    /// GRAMMAR
-
-
-    enum struct production_operator 
-    { and_, or_, some } ;
-
-    template <
-        typename language_t, 
-        auto definition_size_c>
-    struct definition 
-    {
-        using ids_container_t = std::array<language_t, definition_size_c> ;
-        using operator_t = production_operator ;
-        
-        operator_t op ;
-        ids_container_t ids ;
-    } ;
-
-    template <typename language_t>
-    constexpr auto 
-    make_definition (
-        production_operator op, 
-        language_t first_id,
-        auto ... next_id)
-    {
-        using definition_t = definition <language_t, 1 + sizeof...(next_id)> ; 
-        
-        return 
-        definition_t { op, { first_id, next_id... } } ;
-    }
-
-
-
-
-    constexpr auto 
-    and_op (
-        auto first_id, 
-        auto ... next_id)
-    {
-        using po = production_operator ; 
-        
-        return 
-        make_definition (po::and_, first_id, next_id ...) ;
-    }
-
-
-    constexpr auto 
-    or_op (
-        auto first_id, 
-        auto ... next_id)
-    {
-        using po = production_operator ; 
-        
-        return 
-        make_definition (po::or_, first_id, next_id ...) ;
-    }
-        
-        
-    constexpr auto 
-    some_op (
-        auto id)
-    {
-        using po = production_operator ; 
-        
-        return 
-        make_definition (po::some, id) ;
-    }
-
-
-
-
-    template <
-        typename language_t,
-        auto production_size_c> 
-    struct production 
-    {
-        using ids_container_t = std::array<language_t, production_size_c> ;
-        using operator_t = production_operator ;
-
-        language_t id ;
-        operator_t op ;
-        ids_container_t ids ;
-    } ;
-
-
-    template <
-        typename language_t, 
-        auto definition_size_c>
-    constexpr auto 
-    is (
-        language_t id, 
-        definition<language_t, definition_size_c> def)
-    {
-        using production_t = production <language_t, definition_size_c> ;
-
-        return production_t { id, def.op, def.ids } ;
-    }
-
-    
-
-    template <
-        typename language_t, 
-        typename ... production_t>
-    struct productions ;
 }
 
-#define THODD_LANG_OPERATOR_FOR(language_t) \
-template <auto size_c> \
-constexpr auto \
-operator <= (language_t id, thodd::lang::definition<language_t, size_c> def) \
-{ return thodd::lang::is (id, def) ; } \
-\
-\
-constexpr auto \
-operator * (language_t id) \
-{ return thodd::lang::some_op (id) ; } \
-\
-constexpr auto \
-operator > (language_t id, language_t id2) \
-{ return thodd::lang::and_op (id, id2) ; } \
-\
-template<auto size_c>\
-constexpr auto \
-operator > (definition<language_t, size_c> and_def, language_t id2) \
-{ return thodd::lang::and_op (id, id2) ; } \
+
+
 
 
 #endif
